+++ changelog.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Version 1.0.5
+* improved debugging information for FormulaGrader
+* FormulaGrader and IntegralGrader perform whitelist, blacklist, and forbidden_string checks after determining answer correctness. Incorrect answers using forbidden strings / functions are now marked incorrect, while correct answers using forbidden strings / functions raise errors.
+* minor improvements to existing unit tests
+
 ## Version 1.0.4
 * Authors can now specify a custom comparer function for FormulaGrader
 * IntegralGrader now handles complex integrands, and gives meaningful error messages